move sidebarList to a hook
